+++ attributes/default.rb
@@ -3,6 +3,7 @@
 # Cookbook Name:: sidekiq-manager
 # Attributes:: default
 #
-# Copyright 2014, YOUR_COMPANY_NAME
+# Copyright 2014, We're in s.r.o.
 #
 
+default['rails']['applications_root'] = '/u/apps'