+++ metadata.rb
@@ -1,9 +1,9 @@
 name 'sidekiq-manager'
-maintainer 'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
+maintainer "We're in s.r.o."
+maintainer_email 'jiri.kolarik@wereinhq.com'
 license 'All rights reserved'
-description 'Installs/Configures sidekiq-manager'
+description 'Installs/Configures sidekiq-manager with upstart'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version '0.1.0'
 
-depends 'apt'
+depends 'redisio'