+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 This file is used to list changes made in each version of the sidekiq-manager cookbook.
 
 ## 0.1.0
-- [your_name] - Initial release of sidekiq-manager
+- [@jirikolarik] - Initial release of sidekiq-manager